% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_ave.R
\name{gmm_ave}
\alias{gmm_ave}
\title{GMM moments with averaged outcome proxy}
\usage{
gmm_ave(
  theta,
  dat,
  mod = NULL,
  n_y,
  w_idx,
  iv_load_idx,
  iv_reg_idx,
  iv_names_load,
  iv_names_reg,
  y_names
)
}
\description{
GMM moments with averaged outcome proxy
}
