% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_opt.R
\name{gmm_opt}
\alias{gmm_opt}
\title{GMM moments (per-outcome proxy, no inverse-variance aggregation)}
\usage{
gmm_opt(
  theta,
  dat,
  n_y,
  w_idx,
  iv_load_idx,
  iv_reg_idx,
  iv_names_load,
  iv_names_reg,
  y_names
)
}
\description{
GMM moments (per-outcome proxy, no inverse-variance aggregation)
}
