% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_opt_robust.R
\name{gmm_opt_robust}
\alias{gmm_opt_robust}
\title{GMM moments for robust two-step estimation}
\usage{
gmm_opt_robust(
  theta,
  dat,
  n_y,
  w_idx,
  iv_load_idx,
  iv_reg_idx,
  iv_names_load,
  iv_names_reg,
  y_names,
  block = c("joint", "load", "reg"),
  lambda_fix = NULL
)
}
\description{
GMM moments for robust two-step estimation
}
