% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estlatent.R
\name{estlatent}
\alias{estlatent}
\title{Estimate the Causal Effects with Latent Outcomes}
\usage{
estlatent(Z, Y, X = NULL, eta = 1, method = "sem")
}
\arguments{
\item{Z}{a vector of Treatment variable.}

\item{Y}{a data frame that includes the measured outcomes. The outcome in the first column will be set as lambda=1.}

\item{X}{an optional data frame that includes other covariates. The defauly is NULL.}

\item{eta}{tbd}

\item{method}{estimators options.}
}
\description{
Estimate the Causal Effects with Latent Outcomes
}
