% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estlatent.R
\name{estlatent}
\alias{estlatent}
\title{Estimate the Causal Effects with Latent Outcomes}
\usage{
estlatent(Z, Y, X = NULL, eta = 1, method = "sem", IV_Y = T, tau = T)
}
\arguments{
\item{Z}{a vector of Treatment variable.}

\item{Y}{a data frame that includes the measured outcomes. The outcome in the first column will be set as lambda=1.}

\item{X}{an optional data frame that includes other covariates. The defauly is NULL.}

\item{eta}{tbd}

\item{method}{estimators options.}

\item{IV_Y}{Whether use other outcome measures as IVs}

\item{tau}{For GMM only. If not, calculate sample average in two groups}
}
\description{
Estimate the Causal Effects with Latent Outcomes
}
\examples{
\dontrun{
data(test_dat)  # input data
estlatent(test_dat$Z,test_dat[,1:3],X=NULL,eta = 1,method="sem",IV_Y=T,tau=T)
}
}
\references{
Fu, Jiawei, and Donald P. Green. "Causal Inference for Experiments with Latent Outcomes: Key Results and Their Implications for Design and Analysis." (2025).
}
