% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estlatent_robust.R
\name{estlatent_robust}
\alias{estlatent_robust}
\title{Estimate latent-outcome effects (two-step robust GMM with Godambe sandwich)}
\usage{
estlatent_robust(
  mod = NULL,
  Z,
  Y,
  data = NULL,
  method = "sem",
  IV_Y = NULL,
  opt = TRUE
)
}
\arguments{
\item{mod}{Optional regression specification for latent regression.
Recommended form: "~Z+x1+x2". Also accepts "Y~..." or "eta~...".
If NULL, defaults to "~Z".}

\item{Z}{Treatment input: vector/1-col matrix-data.frame, or a single column name when data is provided.}

\item{Y}{Measured outcomes: matrix-data.frame with >=2 columns, or character vector of names when data is provided.}

\item{data}{Optional data frame to resolve character inputs and covariates in mod.}

\item{method}{"sem" or "gmm".}

\item{IV_Y}{Optional loading-IV names for stage-1 loading GMM. Default uses treatment Z only.}

\item{opt}{For each GMM stage: TRUE=two-step efficient GMM, FALSE=one-step GMM.}
}
\description{
Estimate latent-outcome effects (two-step robust GMM with Godambe sandwich)
}
\examples{
\dontrun{
data("test_data", package = "LatentOutcomes")  # input data
m1 <- estlatent_ave(mod="~Z", Z="Z", Y=c("Y1","Y2","Y3"), data=test_dat, method="sem")
summary(m1) # sem model
m2 <- estlatent_ave(mod="~Z", Z="Z", Y=c("Y1","Y2","Y3"), data=test_dat, method="gmm", IV_Y=c("Z","Y2"), opt=TRUE)
summary(m2) # gmm model, using Z and Y2 as IVs for the factor loadings
m3 <- estlatent_ave(mod="~Z+x1+x2", Z="Z", Y=c("Y1","Y2","Y3"), data=test_dat, method="gmm", IV_Y=c("Z","Y1","Y2","Y3"), opt=TRUE) # control covariates x1 and x2, using Z, Y1, Y2, and Y3 as IVs for the factor loadings
summary(m3)
}
}
\references{
Fu, Jiawei, and Donald P. Green. "Causal Inference for Experiments with Latent Outcomes: Key Results and Their Implications for Design and Analysis." (2025).
}
