% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_dat}
\alias{test_dat}
\title{Test dataset}
\format{
A data frame with 1000 observations and 4 variables:
\describe{
\item{Y1}{First outcome measure}
\item{Y2}{Second outcome measure}
\item{Y3}{Third outcome measure}
\item{Z}{Treatment assignment}
}
}
\usage{
test_dat
}
\description{
A dataset
}
\keyword{datasets}
